ajustes estrutura

+++ src/main/java/nature/calculadoraco2/Repositories/AtividadeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface AtividadeRepository extends JpaRepository<Atividade, Integer> {
+public interface AtividadeRepository extends JpaRepository<Atividade, Long> {
 }
